move runtime class to runtime.h
