+++ CMakeLists.txt
@@ -82,7 +82,10 @@ configure_file(
     NEWLINE_STYLE LF)
 
 # installation
-install(FILES "include/aws/lambda-runtime/runtime.h" "include/aws/lambda-runtime/version.h"
+install(FILES "include/aws/http/response.h"
+    DESTINATION "include/aws/http")
+
+install(FILES "include/aws/lambda-runtime/runtime.h" "include/aws/lambda-runtime/version.h" "include/aws/lambda-runtime/outcome.h"
     DESTINATION "include/aws/lambda-runtime")
 
 install(FILES "include/aws/logging/logging.h"
